week3_assignment_modified
