+++ .gitignore
@@ -172,5 +172,4 @@ cython_debug/
 #.idea/
 
 # End of https://www.toptal.com/developers/gitignore/api/django
-
 secrets.json
