+++ posts/urls.py
@@ -3,4 +3,5 @@ from posts.views import *
 
 urlpatterns = [
     path('', hello_world, name = 'hello_world'),
+    path('introduction/', code_reviewer_info, name = 'code_reviewer_info'),
 ]
