+++ posts/views.py
@@ -12,3 +12,25 @@ def hello_world(request):
             'message' : '메시지 전달 성공!',
             'data' : "Hello world",
         })
+# Create your views here.
+
+def code_reviewer_info(request):
+    if request.method == "GET":
+        return JsonResponse({
+            'status' : 200,
+            'success' : True,
+            'message' : '3주차 미션 성공!',
+            'data' : [
+            {
+                "name" : "이기웅",
+                "age" : 24,
+                "major" : "Energy System Engineering"
+
+            },
+            {
+                "name" : "박소은",
+                "age" : 23,
+                "major" : "Computer Science and Engineering"
+            }
+            ]
+        })
