+++ requirements.txt
@@ -1,4 +1,4 @@
 asgiref==3.6.0
 Django==4.1.7
 sqlparse==0.4.3
-tzdata==2023.2
+tzdata==2023.3
